Prototype level view added
Simple level script added to contain collision data.

Level renderer added. This isn't a renderer so much as a manger for
prefab
instantiation of tiles. This is hopelessly inefficient, but should do
for now.

RetardPhysics is a ten-minute attempt at doing some platformer physics.
It contains no code that could be considered actual physics whatsoever,
but is instead a wadge of clumsy trigger response code. This should be
replaced with simple ray-casts to effectively handle collisions other
than down.

I've added a basic tile editor which simply allows painting of
blocked/open cells. This is redundant if we plan to do procedural
levels, but I would imagine those to be a hell of a way down the line.

TODO: Get the level to the AdminView via the wonders of networking and
allow that scene to monitor a connected player's movement.