Network replication changes
The AdminView is currently disabled.  The main menu now loads the
AgentView only.  From the AgentView, you'll either be able to start a
server or connect as a client.

I've modified the replication to use the RigidBody instead of the
transform.  This entailed turning off isKinematic and enabled
interpolation of the position.  The actor is now replicated smoothly,
however the camera follow on the client appears to not use the smoothed
extrapolated position so this seems a bit jerky at the moment.  I'll
look into making smoothing this, but for the time being this is
acceptable.