Character-controller added
This is just a simple ray-cast collider to replace the laughable mess
being used for collisions right now.

It's not ready for consumption yet, but I'm bunging in this early
version to create a history in case I bugger it up during tidying.

This may or may not mangle network behaviour due to no longer involving
a box-collider (though that's due to be reinstated in my next
submission). If so, apologies. I'll sort it if that proves to be the
case.