First test of a simple agent/admin system added.
The new system defines a base MonoBehaviour (WorldObject) for each type
that contains all essential information for a particular item. This base
class also exposes all information to be serialised.

Per-view specialisations are then defined that inherit from

Specific prefabs then inherit from WorldViewObject. These classes handle
all view-specific behaviour and feed the relevant details back into the
WorldObject for synchronisation. This allows the WorldViewObjects to
determine what is updated and when (for cameras, etc).

The current example consists of a door with a sensor in the agent view,
and the same sans-sensor for the admin. I've also added a cheeky network
serialize to give it a go. If one build has the doors set to use the
admin view, and one the agent, both instances can move around the world,
but only one can set the open/closed state of the doors in both. This
isn't represented visually, but the collision updates correctly.

As usual, this is all a bit ghetto and poorly named.